refact: use same strategy to get metadata
